scaffoldState wrapper
