+++ Jetsnack/app/src/main/java/com/example/jetsnack/ui/components/Scaffold.kt
@@ -16,6 +16,7 @@
 
 package com.example.jetsnack.ui.components
 
+import android.content.res.Resources
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.material.DrawerDefaults
@@ -27,14 +28,23 @@ import androidx.compose.material.SnackbarHost
 import androidx.compose.material.SnackbarHostState
 import androidx.compose.material.rememberScaffoldState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.ReadOnlyComposable
+import androidx.compose.runtime.Stable
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.Dp
+import com.example.jetsnack.model.SnackbarManager
 import com.example.jetsnack.ui.theme.JetsnackTheme
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.launch
 
 /**
- * Wrap Material [androidx.compose.material.Scaffold] and set [JetsnackTheme] colors.
+ * Wrap Material [androidx.compose.material.Scaffold] and set [JetsnackTheme] colors and [JetsnackScaffoldState].
  */
 @Composable
 fun JetsnackScaffold(
@@ -75,4 +85,56 @@ fun JetsnackScaffold(
         contentColor = contentColor,
         content = content
     )
+}
+
+/**
+ * Remember and created an instance of [JetsnackScaffoldState]
+ */
+@Composable
+fun rememberJetsnackScaffoldState(
+    scaffoldState: ScaffoldState = rememberScaffoldState(),
+    snackbarManager: SnackbarManager = SnackbarManager,
+    resources: Resources = resources(),
+    coroutineScope: CoroutineScope = rememberCoroutineScope()
+): JetsnackScaffoldState = remember(scaffoldState, snackbarManager, resources, coroutineScope) {
+    JetsnackScaffoldState(scaffoldState, snackbarManager, resources, coroutineScope)
+}
+
+/**
+ * Responsible for holding [ScaffoldState], handles the logic of showing snackbar messages
+ */
+@Stable
+class JetsnackScaffoldState(
+    val scaffoldState: ScaffoldState,
+    private val snackbarManager: SnackbarManager,
+    private val resources: Resources,
+    coroutineScope: CoroutineScope
+) {
+    // Process snackbars coming from SnackbarManager
+    init {
+        coroutineScope.launch {
+            snackbarManager.messages.collect { currentMessages ->
+                if (currentMessages.isNotEmpty()) {
+                    val message = currentMessages[0]
+                    val text = resources.getText(message.messageId)
+                    // Notify the SnackbarManager so it can remove the current message from the list
+                    snackbarManager.setMessageShown(message.id)
+                    // Display the snackbar on the screen. `showSnackbar` is a function
+                    // that suspends until the snackbar disappears from the screen
+                    scaffoldState.snackbarHostState.showSnackbar(text.toString())
+                }
+            }
+        }
+    }
+}
+
+/**
+ * A composable function that returns the [Resources]. It will be recomposed when `Configuration`
+ * gets updated.
+ */
+@Composable
+@ReadOnlyComposable
+private fun resources(): Resources {
+    LocalConfiguration.current
+    return LocalContext.current.resources
 }
